Extract helper functions in the TreeView into a separate file
